backend/foodgram/requirements.txt
asgiref==3.5.0
Django==4.0.3
django-colorfield==0.7.2
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
djoser==2.1.0
drf-extra-fields==3.4.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
python-dotenv==0.20.0
PyJWT==2.3.0
pytz==2020.1
requests==2.27.1
typing_extensions==4.2.0